+++ tests/test_hooks.py
@@ -2,7 +2,7 @@ import os
 from io import BytesIO
 from unittest.mock import MagicMock, patch
 
-from isort import hooks
+from isort import exceptions, hooks
 
 
 def test_git_hook(src_dir):
@@ -25,4 +25,16 @@ def test_git_hook(src_dir):
 
         with patch("subprocess.run", MagicMock(return_value=FakeProcessResponse())) as run_mock:
             with patch("isort.api", MagicMock(return_value=False)):
+                hooks.git_hook(modify=True)
+
+    # Test with skipped file returned from git
+    with patch(
+        "isort.hooks.get_lines", MagicMock(return_value=[os.path.join(src_dir, "main.py")])
+    ) as run_mock:
+
+        class FakeProcessResponse(object):
+            stdout = b"# isort: skip-file\nimport b\nimport a\n"
+
+        with patch("subprocess.run", MagicMock(return_value=FakeProcessResponse())) as run_mock:
+            with patch("isort.api", MagicMock(side_effect=exceptions.FileSkipped("", ""))):
                 hooks.git_hook(modify=True)