+++ tests/test_ticketed_features.py
@@ -19,3 +19,16 @@ def test_thing(): pass
 """,
         show_diff=True,
     )
+
+
+def test_isort_automatically_removes_duplicate_aliases_issue_1193():
+    """Test to ensure isort can automatically remove duplicate aliases.
+    See: https://github.com/timothycrosley/isort/issues/1281
+    """
+    assert isort.check_code("from urllib import parse as parse\n", show_diff=True)
+    assert (
+        isort.code("from urllib import parse as parse", remove_redundant_aliases=True)
+        == "from urllib import parse\n"
+    )
+    assert isort.check_code("import os as os\n", show_diff=True)
+    assert isort.code("import os as os", remove_redundant_aliases=True) == "import os\n"