+++ CHANGELOG.md
@@ -5,8 +5,10 @@ NOTE: isort follows the [semver](https://semver.org/) versioning standard.
 
 ### 5.1.0 July TBD, 2020
   - isort now throws an exception if an invalid settings path is given (issue #1174).
+  - Implemented support for automatic redundant alias removal (issue #1281).
   - Fixed #1178: support for semicolons in decorators.
   - Fixed #1315: Extra newline before comment with -n + --fss.
+
 **Formatting changes implied:**
   - Fixed #1280: rewrite of as imports changes the behavior of the imports.